requirements.txt
langgraph>=0.0.21
langchain-openai>=0.0.8
langchain-core>=0.1.33
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
numpy>=1.24.0
scipy>=1.11.0
opencv-python>=4.8.0
structlog>=24.1.0
websockets>=12.0
requests>=2.31.0
ollama>=0.4.0
anthropic>=0.40.0
google-genai>=1.0.0
pytest-asyncio>=0.23.0
pyyaml>=6.0
shapely>=2.0
